+++ Changelog.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+* Fixed a bug that caused missing graphs in some cases when multiple iterations
+  of a benchmark are present in the bechmark results file.
+
+* Better error reporting to pinpoint errors when a problem occurs.
+
 ## 0.1.1
 
 * Support GHC 8.4